Move interfaces to Chirp.Core
As per request of TA, I have moved the interfaces to Core. Interfaces found within the class that implements them have also been moved out to their own class.
