+++ src/Chirp.Core/RepositoryInterfaces/ICheepRepository.cs
@@ -1,6 +1,6 @@
 ﻿using Chirp.Core.DomainModel;
 
-namespace Chirp.Infrastructure.Repositories;
+namespace Chirp.Core.RepositoryInterfaces;
 
 public interface ICheepRepository
 {
@@ -8,7 +8,6 @@ public interface ICheepRepository
     public Task<List<Cheep>> ReadCheepsFrom(int page, int authorId);
     public Task<int> GetTotalCheeps();
     public Task<int> GetTotalCheepsFor(int authorId);
-    
     public Task CreateCheep(CheepDTO newCheep);
     public Task AddCheep(Cheep cheep);
 }
