+++ src/Chirp.Core/ServiceInterfaces/ICheepService.cs
@@ -1,0 +1,8 @@
+namespace Chirp.Core.ServiceInterfaces;
+
+public interface ICheepService
+{
+    int CurrentPage { get; set; }
+    public Task<List<CheepDTO>> GetCheeps();
+    public Task<int> GetTotalCheeps();
+}
