+++ src/Chirp.Infrastructure/Repositories/CheepRepository.cs
@@ -1,4 +1,6 @@
-﻿using Chirp.Core.DomainModel;
+﻿using Chirp.Core;
+using Chirp.Core.DomainModel;
+using Chirp.Core.RepositoryInterfaces;
 using Chirp.Infrastructure.Database;
 
 using Microsoft.EntityFrameworkCore;
@@ -9,6 +11,7 @@ public class CheepRepository : ICheepRepository
 {
     private const int PAGE_SIZE = 32;
     private readonly ChirpDbContext _dbContext;
+
     public CheepRepository(ChirpDbContext dbContext)
     {
         _dbContext = dbContext;
@@ -35,7 +38,7 @@ public class CheepRepository : ICheepRepository
     public async Task<List<Cheep>> ReadCheepsFrom(int page, int authorId)
     {
         if (page < 1) throw new ArgumentOutOfRangeException(nameof(page));
-        
+
         var query = await _dbContext.Cheeps
             .Where(cheep => cheep.AuthorId == authorId)
             .Include(cheep => cheep.Author)
@@ -50,12 +53,12 @@ public class CheepRepository : ICheepRepository
     {
         return _dbContext.Cheeps.CountAsync();
     }
-    
+
     public Task<int> GetTotalCheepsFor(int authorId)
     {
         return _dbContext.Cheeps.CountAsync(cheep => cheep.AuthorId == authorId);
     }
-    
+
     /// <summary>
     /// creates new cheep messages, checks if user is logged in, identified by unique email
     /// </summary>
@@ -64,24 +67,22 @@ public class CheepRepository : ICheepRepository
     /// <exception cref="Exception"> Is thrown if the user does not exist as an author in the database,
     /// in the future will be redirected to loggin/create-user page
     /// </exception>
-    public async Task CreateCheep(CheepDTO newCheep) {
+    public async Task CreateCheep(CheepDTO newCheep)
+    {
         var command = await _dbContext.Authors.SingleOrDefaultAsync(a => a.Name == newCheep.Author);
         if (command == null)
         {
             throw new Exception("Author does not exist! Create a new author before you can write cheeps to timeline.");
         }
-        
+
         var cheep = new Cheep()
         {
-            AuthorId = command.AuthorId,
-            Author = command,
-            Text = newCheep.Message, 
-            TimeStamp = DateTime.UtcNow,
+            AuthorId = command.AuthorId, Author = command, Text = newCheep.Message, TimeStamp = DateTime.UtcNow,
         };
         _dbContext.Cheeps.Add(cheep);
         await _dbContext.SaveChangesAsync();
     }
-    
+
     /// <summary>
     /// Adds a new cheep to the database.
     /// </summary>
@@ -91,6 +92,4 @@ public class CheepRepository : ICheepRepository
         _dbContext.Cheeps.Add(cheep);
         await _dbContext.SaveChangesAsync();
     }
-
 }
-
